+++ dragonflybsd/DragonFlyBSDProcess.c
@@ -86,7 +86,7 @@ static void DragonFlyBSDProcess_writeField(const Process* this, RichString* str,
    RichString_appendWide(str, attr, buffer);
 }
 
-static long DragonFlyBSDProcess_compareByKey(const Process* v1, const Process* v2, ProcessField key) {
+static int DragonFlyBSDProcess_compareByKey(const Process* v1, const Process* v2, ProcessField key) {
    const DragonFlyBSDProcess* p1 = (const DragonFlyBSDProcess*)v1;
    const DragonFlyBSDProcess* p2 = (const DragonFlyBSDProcess*)v2;
